Do not swallow fatal errors (e.g. MEMORY_LIMIT_EXCEEDED) in tryDeserialize
The serialization "try-pattern" (`tryDeserializeText`/`tryRead` and the
restore-on-failure paths) catches every exception and returns `false`, i.e.
treats the value as "did not parse". That is correct for genuine parse errors,
but it also swallows fatal errors such as `MEMORY_LIMIT_EXCEEDED` (and query
cancellation, logical errors, ...): during best-effort parsing
(`input_format_null_as_default`, `input_format_allow_errors_num`/`ratio`,
schema inference) such an error is silently turned into a default / skipped row
instead of aborting the query.

Make the try-pattern distinguish the two:
- Move `isParseError` out of `Processors/Formats/IRowInputFormat` into a shared
  `Formats/ParseError.h` so the serialization layer (which is below Processors)
  can use it; add `rethrowIfNotParseError`, which rethrows the currently handled
  exception unless it is a parse error.
- Call it from every `tryDeserialize` `catch (...)` site: the generic
  `SimpleTextSerialization` base and the `ISerialization` helper, the leaf
  `String`/`FixedString`/`CustomSimpleText` patterns, and the composite restores
  (`SerializationTuple::addElementSafe`, `Array`, `Map`, `Nullable`, `QBit`).

Now only a genuine parse error is reported as a not-parsed value; every other
error propagates. The set of parse errors is the same allowlist
`IRowInputFormat` already used to decide which rows may be skipped, so the
row-input error-skipping behaviour is unchanged.
