+++ src/DataTypes/Serializations/SerializationCustomSimpleText.cpp
@@ -1,5 +1,6 @@
 #include <DataTypes/Serializations/SerializationCustomSimpleText.h>
 
+#include <Formats/ParseError.h>
 #include <IO/ReadBufferFromString.h>
 #include <IO/ReadHelpers.h>
 #include <IO/WriteBufferFromString.h>
@@ -49,6 +50,8 @@ bool SerializationCustomSimpleText::tryDeserializeText(DB::IColumn & column, DB:
     }
     catch (...) // Ok: tryDeserializeText is a try-pattern
     {
+        /// Other errors (e.g. MEMORY_LIMIT_EXCEEDED) must propagate, not be reported as a failed parse.
+        rethrowIfNotParseError();
         return false;
     }
 }
